API: Compute mpf_lwkj with image and no Phantom
Updating the mtf_lwkj method such that it is computed without requiring
a Phantom object as a template. This allows this function to be used on
real data with less work. There are new keywords for this function
because information needs to be supplied that was taken from the
Phantom. The frequency space is now sampled linearly.

The demo notebook has been updated with the new API.